Update a lot of things.

+++ twilio-prototype/src/client/javascript/twilio_admin.js
@@ -20,20 +20,46 @@ app.config(['$routeProvider', '$locationProvider',
     });
 }]);
 
+// This controls the user display page, grab the users and their messages
 app.controller('UserController', ['$scope', '$routeParams', '$location', '$http',
   function($scope, $routeParams, $location, $http) {
+    $scope.updating = false;
     $scope.messages = [];
     $scope.user = {};
     $http.get('/api/u/id/' + $routeParams.id).success(function(data) {
       $scope.user = data;
+      $scope.originaluser = data;
     });
 
     $http.get('/api/u/id/' + $routeParams.id + '/m').success(function(data) {
       $scope.messages = data;
     });
+
+    // Save the new user information and send it to the server via a put method
+    $scope.saveUser = function() {
+      var c = confirm('Do you wish to update this user\'s info?');
+
+      if(c === true) {
+        var options = {
+          headers: { 'Content-Type': 'application/x-www-form-urlencoded' }
+        };
+        console.log('Push new info to server!');
+        $http.put('/api/u', transformRequest($scope.user), options).success(function(data) {
+          console.log(data);
+        });
+        $scope.updating = false;
+      } else {
+        $scope.updating = false;
+        $http.get('/api/u/id/' + $routeParams.id).success(function(data) {
+          $scope.user = data;
+          $scope.originaluser = data;
+        });
+      }
+    };
   }
 ]);
 
+// This controls the main admin screen where all the action happens
 app.controller('AdminController', ['$scope', '$http', '$location',
   function($scope, $http, $location) {
     // Grab the admins info for funsies
@@ -51,6 +77,7 @@ app.controller('AdminController', ['$scope', '$http', '$location',
 
     $scope.add_report_string = 'Add Reporting Period';
 
+    // Show/hide the reporting period addition form
     $scope.toggle_show_reporting = function() {
       $scope.show_reporting = !$scope.show_reporting;
       if($scope.show_reporting) {
@@ -62,7 +89,9 @@ app.controller('AdminController', ['$scope', '$http', '$location',
       }
     };
 
+    // Add the reporting period to the database based on the current user input
     $scope.start_reporting = function() {
+      // For convience check and output the current date
       console.log(Date.now());
       if ($scope.newreport.startDate < Date.now() && $scope.newreport.endDate > Date.now()) {
         console.log('Now falls between that range');
@@ -70,6 +99,8 @@ app.controller('AdminController', ['$scope', '$http', '$location',
         console.log('You are not in range!');
       }
 
+      // Because we have to user form urlencoded for twilio we must convert our json
+      // objects into the type. Grab the information and convert then post
       var options = {
         headers: { 'Content-Type': 'application/x-www-form-urlencoded' }
       };
@@ -94,6 +125,7 @@ app.controller('AdminController', ['$scope', '$http', '$location',
       if($scope.show_periods) {
         $scope.show_periods_string = 'Hide Reporting Periods';
         $http.get('/api/p').success(function(data) {
+          $scope.periods = [];
           $scope.periods = data;
         });
       } else {
@@ -102,13 +134,16 @@ app.controller('AdminController', ['$scope', '$http', '$location',
     };
 
     $scope.delete_period = function(index) {
-      $http.delete('/api/p/id/' + $scope.periods[index]._id).success(function(data) {
-        console.log(data);
+      var c = confirm('Are you sure you want to delete this reporting period?');
+      if(c === true) {
+        $http.delete('/api/p/id/' + $scope.periods[index]._id).success(function(data) {
+          console.log(data);
 
-        $http.get('/api/p').success(function(data) {
-          $scope.periods = data;
+          $http.get('/api/p').success(function(data) {
+            $scope.periods = data;
+          });
         });
-      });
+      }
     };
 
     // USER LIST
@@ -136,7 +171,10 @@ app.controller('AdminController', ['$scope', '$http', '$location',
     $scope.newuser = {
       firstname: '',
       lastname: '',
-      number: ''
+      number: '',
+      salutation: '',
+      farmerid: '',
+      contractType: ''
     };
 
     // code to call when clicking add cell button
@@ -168,6 +206,9 @@ app.controller('AdminController', ['$scope', '$http', '$location',
         $scope.newuser.firstname = '';
         $scope.newuser.lastname = '';
         $scope.newuser.number = '';
+        $scope.newuser.salutation = '';
+        $scope.newuser.contractType = '';
+        $scope.newuser.farmerid = '';
       });
     };
 
@@ -203,20 +244,25 @@ app.controller('AdminController', ['$scope', '$http', '$location',
 
     // Delete a user from the user list
     $scope.delete_user = function(index) {
-      $http.delete('/api/u/id/' + $scope.users[index]._id).success(function(data) {
-        console.log(data);
+      var c = confirm('Are you sure you want to delete this user?');
+      if(c === true) {
+        $http.delete('/api/u/id/' + $scope.users[index]._id).success(function(data) {
+          console.log(data);
 
-        $http.get('/api/u').success(function(data) {
-          $scope.users = data;
+          $http.get('/api/u').success(function(data) {
+            $scope.users = data;
+          });
         });
-      });
+      }
     };
   }
 ]);
 
+// Show the current period information
 app.controller('PeriodController', ['$scope', '$http', '$location', '$routeParams',
   function($scope, $http, $location, $routeParams) {
     $scope.messages = [];
+    $scope.users = [];
     $scope.period = '';
     $scope.exportArray = [];
 
@@ -227,9 +273,13 @@ app.controller('PeriodController', ['$scope', '$http', '$location', '$routeParam
 
     $http.get('/api/p/id/' + $routeParams.id + '/m').success(function(data) {
       for(var i = 0; i < data.length; i++) {
-        $scope.exportArray.push({a: data[i].sender, b: data[i].body});
+        $scope.exportArray.push({ a: data[i].sender, b: data[i].body });
       }
       $scope.messages = data;
+
+      $http.get('/api/p/id/' + $routeParams.id + '/u').success(function(users) {
+        $scope.users = users;
+      });
     });
   }
 ]);
